pyproject.toml
[tool.black]
line-length = 120
target-version = ['py36']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.pytest_cache
  | \.tox
  | \.venv
  | venv
  | build
  | dist
  | setup.py
)/
'''

[tool.poetry]
name = "mqttlistener"
version = "0.1.0"
description = "CLI MQTT Listener Template"
authors = ["Zobayer Hasan <zobayer1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
mqttlistener = "mqttlistener.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
click = "^8.1.7"
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
